+++ COMP371_RaytracerBase/code/src/Light.h
@@ -6,13 +6,13 @@ class Light
 {
 public:
     string type;
-    Vector3f centre;
-    Vector3f id;
-    Vector3f is;
+    Vector3d centre;
+    Vector3d id;
+    Vector3d is;
     int n = 1;
     bool usecenter = true;
     bool visible = true;
 
-    Light(string _type, Vector3f _centre, Vector3f _id, Vector3f _is)
+    Light(string _type, Vector3d _centre, Vector3d _id, Vector3d _is)
     : type(_type), centre(_centre), id(_id), is(_is) {}
 };
